+++ ChangeLog.md
@@ -1,3 +1,11 @@
+## v0.1 - 2024-04-23
+
+### Added
+
+- Add database connection
+- Add DbPdo entity
+- Add Session entity/functions
+
 ## v0.0 - 2024-04-23
 
 ### Added